pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gdown>=5.2.0",
    "numpy>=2.2.4",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "tensorflow>=2.14.0",
    "trafilatura>=2.0.0",
]
